[front] chore: Migrate productboard MCP server
